+++ 2018.1/atividades/03-SimpleInterpreter/ast-simple-interpreter/src/br/ufpe/cin/if688/symboltable/IntAndTable.java
@@ -2,9 +2,9 @@ package br.ufpe.cin.if688.symboltable;
 
 
 public class IntAndTable {
-	public int result;
+	public double result;
 	public Table table;
-	public IntAndTable(int res, Table t){
+	public IntAndTable(double res, Table t){
 		result = res;
 		table = t;
 	}
