+++ 2018.1/atividades/03-SimpleInterpreter/ast-simple-interpreter/src/br/ufpe/cin/if688/symboltable/Table.java
@@ -3,10 +3,10 @@ package br.ufpe.cin.if688.symboltable;
 public class Table {
 
 	public String id;
-	public int value;
+	public double value;
 	public Table tail;
 
-	public Table(String i, int v, Table t) {
+	public Table(String i, double v, Table t) {
 		id = i;
 		value = v;
 		tail = t;
